backend_fastapi/requirements.txt
# Disaster Resilience AI — Backend Dependencies
# Install with: pip install -r requirements.txt

fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
pydantic[email]>=2.6.0
numpy>=1.26.0
scikit-learn>=1.4.0

# Database
supabase>=2.0.0
python-dotenv>=1.0.0
